# shared-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Melanie Data API '
  description: 'A high performance & centrally cached API service for premium bots. '
  version: ' commit e2ffc88 ok @ Sep 10 11:37 pm'
servers:
- url: https://dev.melaniebot.net
  description: montreal
paths:
  /api/ai/nsfw_check:
    get:
      tags:
      - ai
      summary: Evaluate Image Safety
      description: Returns probabilities of the image containing racy or adult content
      operationId: evaluateImageSafety
      parameters:
      - required: true
        schema:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
        name: url
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageEvaluationResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/ocr:
    post:
      tags:
      - ai
      summary: Read Text From Image
      description: Perform OCR on a text image. Responses may take up to 10 seconds
        to return!
      operationId: readTextfromImage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OCRRquest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OCRReadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/avatar:
    get:
      tags:
      - ai
      summary: Create Pixel Avatar
      description: Create a colorful pixel art totally avatar
      operationId: createpixelavatar
      parameters:
      - description: AI generation prompt
        required: true
        schema:
          type: string
          maxLength: 500
          title: Idea
          description: AI generation prompt
        name: idea
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIImageGenerationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/cyberpunk:
    get:
      tags:
      - ai
      summary: Create Cyberpunk Avatar
      description: Create a Cyberpunk avatar
      operationId: createcyberpunkavatar
      parameters:
      - description: AI generation prompt
        required: true
        schema:
          type: string
          maxLength: 500
          title: Idea
          description: AI generation prompt
        name: idea
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIImageGenerationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/creative:
    get:
      tags:
      - ai
      summary: Create Fantasy Art Avatar
      description: Creative image generator
      operationId: createfantasyartavatar
      parameters:
      - description: AI generation prompt
        required: true
        schema:
          type: string
          maxLength: 500
          title: Idea
          description: AI generation prompt
        name: idea
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIImageGenerationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/dalle-2:
    get:
      tags:
      - ai
      summary: Create Art With Dalle-2
      description: Creative dalle-2 generator
      operationId: createartwithdalle2
      parameters:
      - description: AI generation prompt
        required: true
        schema:
          type: string
          maxLength: 500
          title: Idea
          description: AI generation prompt
        name: idea
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIImageGenerationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ai/segment_bg:
    get:
      tags:
      - ai
      summary: Remove Backgrounds From Image
      description: Performs AI based image segmentation that is more powerful than
        traditional rembg tooling.
      operationId: removebackgroundsfromimage
      parameters:
      - required: true
        schema:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
        name: url
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            image/png: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AIImageGenerationResponse:
      properties:
        url:
          type: string
          title: Url
          description: URL of the image generated. Image should be immediately avaliable.
        idea:
          type: string
          title: Idea
      type: object
      required:
      - url
      - idea
      title: AIImageGenerationResponse
...
    AdvancedInfo:
      properties:
        key:
          type: string
          title: Key
        value:
          type: integer
          title: Value
      type: object
      title: AdvancedInfo
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ImageEvaluationResult:
      properties:
        adult:
          type: boolean
          title: Adult
          description: Indicates if an image is classified as adult.
        racy:
          type: boolean
          title: Racy
          description: Indicates if the image is classified as racy.
        adult_classification_score:
          type: number
          title: Adult Classification Score
          description: Probability image is adult.
        racy_classification_score:
          type: number
          title: Racy Classification Score
          description: Probability image is racy.
        advanced_info:
          items:
            $ref: '#/components/schemas/AdvancedInfo'
          type: array
          title: Advanced Info
      type: object
      title: ImageEvaluationResult
...
    OCRReadResponse:
      properties:
        display_text:
          type: string
          title: Display Text
        lines:
          items:
            type: string
          type: array
          title: Lines
          default: []
      type: object
      title: OCRReadResponse
    OCRRquest:
      properties:
        url:
          type: string
          title: Url
          description: URL of the image to be read. PNG or JPEG supported. Must be
            larger than 50x50
      type: object
      required:
      - url
      title: OCRRquest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...